fix(test): move syscall.Chflags test to _darwin_test.go for Linux CI compat
